requirements.txt
# SpectaSyncAI — Backend Requirements
# @07_modern_polyglot_standards compliant | audited 2026-04-16
# Runtime: Python 3.12+ | google-adk 1.0.0 GA

# ── Google AI / ADK ────────────────────────────────────────────────────────────
google-adk>=1.0.0
google-genai>=1.10.0                  # replaces deprecated google-generativeai
google-cloud-aiplatform>=1.93.0

# ── API / Server ───────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# ── Data Validation ────────────────────────────────────────────────────────────
pydantic>=2.11.0
pydantic-settings>=2.8.0

# ── MCP / Toolbox ──────────────────────────────────────────────────────────────
mcp[server]>=1.9.0

# ── Database / Vector ──────────────────────────────────────────────────────────
asyncpg>=0.30.0
pgvector>=0.3.6

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv>=1.1.0
httpx>=0.28.0

# ── Testing ────────────────────────────────────────────────────────────────────
pytest>=8.3.0
pytest-asyncio>=0.24.0
